+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 
 -------------------------
 
+## 2021-03-10
+
+- REL-518415 - New version 8.1.1.1 created - Relativity.API added as a dependency, GetDbContext() set to throw NotImplementedException().
+
+-------------------------
+
 ## 2021-03-09
 
 - REL-518415 - New version 8.1.0.1 created (RSAPI removal and refactor)